CBP - stock Transactions - Crypto Money Restock - QA #3148 Fix Implementation BigDecimal

+++ CBP/plugin/stock_transactions/fermat-cbp-plugin-stock-transactions-crypto-money-restock-bitdubai/src/main/java/com/bitdubai/fermat_cbp_plugin/layer/stock_transactions/crypto_money_restock/developer/bitdubai/version_1/structure/CryptoMoneyRestockTransactionImpl.java
@@ -5,6 +5,7 @@ import com.bitdubai.fermat_cbp_api.all_definition.business_transaction.CryptoMon
 import com.bitdubai.fermat_cbp_api.all_definition.enums.OriginTransaction;
 import com.bitdubai.fermat_cbp_api.all_definition.enums.TransactionStatusRestockDestock;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.util.UUID;
 
@@ -19,10 +20,10 @@ public class CryptoMoneyRestockTransactionImpl implements CryptoMoneyTransaction
     String                          cryWalletPublicKey;
     String                          memo;
     String                          concept;
-    float                           amount;
+    BigDecimal                      amount;
     Timestamp                       timeStamp;
     TransactionStatusRestockDestock transactionStatus;
-    float                           priceReference;
+    BigDecimal                      priceReference;
     OriginTransaction originTransaction;
 
     public CryptoMoneyRestockTransactionImpl(){
@@ -36,10 +37,10 @@ public class CryptoMoneyRestockTransactionImpl implements CryptoMoneyTransaction
                                              String cryWalletPublicKey,
                                              String memo,
                                              String concept,
-                                             float amount,
+                                             BigDecimal amount,
                                              Timestamp timeStamp,
                                              TransactionStatusRestockDestock transactionStatus,
-                                             float priceReference,
+                                             BigDecimal priceReference,
                                              OriginTransaction originTransaction){
         this.transactionId      = transactionId;
         this.actorPublicKey     = actorPublicKey;
@@ -118,12 +119,12 @@ public class CryptoMoneyRestockTransactionImpl implements CryptoMoneyTransaction
     }
 
     @Override
-    public float getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
     @Override
-    public void setAmount(float amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
 
@@ -158,12 +159,12 @@ public class CryptoMoneyRestockTransactionImpl implements CryptoMoneyTransaction
     }
 
     @Override
-    public float getPriceReference() {
+    public BigDecimal getPriceReference() {
         return priceReference;
     }
 
     @Override
-    public void setPriceReference(float priceReference) {
+    public void setPriceReference(BigDecimal priceReference) {
         this.priceReference = priceReference;
     }
